requirements.txt
streamlit
pandas
google-cloud-vision
google-auth
google-api-python-client